requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
gunicorn==21.2.0
sqlalchemy>=1.4.0,<1.5.0
pydantic>=1.8.0,<2.0.0
pydantic-settings==2.1.0
pydantic[email]==2.6.1
passlib[bcrypt]>=1.7.4,<1.8.0
python-jose[cryptography]>=3.3.0,<3.4.0
python-multipart>=0.0.5,<0.0.6
bcrypt==4.1.2
python-dotenv==1.0.1
email-validator==2.1.0.post1
alembic==1.13.1
python-logging-loki==0.3.1
psycopg2-binary==2.9.9 